+++ pages/custom_blocks.py
@@ -1,0 +1,72 @@
+from django import forms
+
+from wagtail.core import blocks
+from wagtail.core.blocks import FieldBlock, StructBlock
+from wagtail.images.blocks import ImageChooserBlock
+from wagtail.documents.blocks import DocumentChooserBlock
+
+
+
+class ImageFormatChoiceBlock(FieldBlock):
+    field = forms.ChoiceField(choices=(
+        ('left', 'Wrap left'), ('right', 'Wrap right'), ('mid', 'Mid width'), ('full', 'Full width'),))
+
+
+class ImageBlock(StructBlock):
+    image = ImageChooserBlock(required=False)
+    alt_text = blocks.CharBlock(required=False)
+    link = blocks.URLBlock(required=False)
+    alignment = ImageFormatChoiceBlock()
+    identifier = blocks.CharBlock(required=False, help_text="Used by the frontend for Google Analytics.")
+
+
+class APIImageChooserBlock(ImageChooserBlock): # Use this block to return the path in the page API, does not support alt_text and alignment
+    def get_api_representation(self, value, context=None):
+        try:
+            return ImageSerializer(context=context).to_representation(value)
+        except AttributeError:
+            return None
+
+class ColumnBlock(blocks.StructBlock):
+    heading = blocks.CharBlock(required=False)
+    content = blocks.RichTextBlock(required=False)
+    image = ImageBlock(required=False, help_text='Callout boxes 940x400, Home page boxes 1464x640')
+    document = DocumentChooserBlock(required=False)
+    cta = blocks.CharBlock(required=False)
+    link = blocks.URLBlock(required=False)
+
+    class Meta:
+        icon = 'placeholder'
+
+
+class FAQBlock(blocks.StructBlock):
+    question = blocks.RichTextBlock(required=True)
+    slug = blocks.CharBlock(required=True)
+    answer = blocks.RichTextBlock(required=True)
+    document = DocumentChooserBlock(required=False)
+
+    class Meta:
+        icon = 'placeholder'
+
+
+class BookProviderBlock(blocks.StructBlock):
+    name = blocks.CharBlock()
+    blurb = blocks.TextBlock(required=False)
+    icon = ImageChooserBlock()
+    cta = blocks.CharBlock()
+    url = blocks.URLBlock()
+    canadian = blocks.BooleanBlock(required=False)
+
+    class Meta:
+        icon = 'document'
+
+    def get_api_representation(self, value, context=None):
+        if value:
+            return {
+                'name': value['name'],
+                'blurb': value['blurb'],
+                'icon': build_image_url(value['icon']),
+                'cta': value['cta'],
+                'url': value['url'],
+                'canadian': value['canadian']
+            }
